==> 2_cleaning-analysis.ipynb <==
# Cleaning and analyzing - contract processes
This notebook will be used for cleaning and analyzing the data scraped from [Sercop's search page](https://modulocomprascorporativas.compraspublicas.gob.ec/ProcesoContratacion/compras/PC/buscarProceso.cpe?sg=1#), scraped with the previuos notebook. 

import pandas as pd

df = pd.read_csv("data/scraped_results_2025-07-09_12-47.csv")

df.info()

df = df.drop(columns="semestre")

df[df["objeto"].str.contains('"')]

df["provincia"] = df["ubicacion"].str.split("/").str[0].str.lower().str.strip()
df["canton"] = df["ubicacion"].str.split("/").str[1].str.lower().str.strip()

df = df.drop(columns="ubicacion")

df["objeto"] = (
    df["objeto"].str.replace("\\", "", regex=False)
        .str.replace('"', '')
        .str.replace("TICA : ", "TICA: ")
        .str.replace("TRACIÓN : ", "TRACIÓN: ")
        .str.replace("MERCIAL : ", "MERCIAL: ")
        .str.replace("-CONCEN", "- CONCEN")
        .str.replace("-PRESEN", "- PRESEN")
        .str.replace(r"\s+", " ", regex=True)
        .str.strip()
)

df["presupuesto"] = df["presupuesto"].str.replace("$", "").str.replace(",", "").astype(float)

df["principio_activo"] = (df["objeto"]
    .str.split("DCI: ").str[1]
    .str.split("- FOR").str[0]
    .str.lower()
    .str.strip()
                         )

df["forma_farmaceutica"] = (df["objeto"]
    .str.split("TICA:").str[1]
    .str.split("- CON").str[0]
    .str.lower()
    .str.strip()
                           )

df["concentracion"] = (df["objeto"]
    .str.split("TRACIÓN:").str[1]
    .str.split("- PRE").str[0]
    .str.lower()
    .str.strip()
                      )

df["presentacion_comercial"] = (df["objeto"]
    .str.split("CIAL:").str[1]
    .str.lower()
    .str.strip()
                               )

df = df.rename(columns={"objeto": "item"})

df.columns

df["forma_farmaceutica"] = (df["forma_farmaceutica"]
    .str.replace("-", "")
    .str.replace("oralpolvo", "oral polvo")
    .str.replace(r"\s+", " ", regex=True)
                           )

df["principio_activo"] = (df["principio_activo"]
    .str.replace("-", "")
    .str.replace("/", "")
    .str.replace(r"\s+", " ", regex=True)
                         )

df = df[['codigo', 'detalle_url', 'item', 'principio_activo', 'forma_farmaceutica', 'concentracion',
         'presentacion_comercial', 'estado', 'presupuesto', 'provincia', 'canton', 'fecha_publicacion']]

df.head(1)

df.info()

# Source: https://compracorporativasalud.compraspublicas.gob.ec/SICM/?page_id=184
portfolio = pd.read_excel("PORTAFOLIO-DE-PUBLICACIÓN-CONSOLIDADO-15-08-2022.xlsx", skiprows=3)

portfolio["NOMBRE GENERICO"] = (portfolio["NOMBRE GENERICO"]
    .str.replace("(", "")
    .str.replace(")", "")
    .str.lower()
    .str.strip()
                               )

portfolio["DESCRIPCIÓN DEL GRUPO ANATÓMICO"] = (portfolio["DESCRIPCIÓN DEL GRUPO ANATÓMICO"]
    .str.lower()
    .str.strip()
                               )

portfolio_clean = portfolio.drop_duplicates(subset=["NOMBRE GENERICO"])

df = pd.merge(
    df, 
    portfolio_clean[["NOMBRE GENERICO", "DESCRIPCIÓN DEL GRUPO ANATÓMICO"]], 
    left_on="principio_activo", 
    right_on="NOMBRE GENERICO", 
    how="left"
).drop(columns=["NOMBRE GENERICO"])

df = df.rename(columns={"DESCRIPCIÓN DEL GRUPO ANATÓMICO": "descripcion"})

df.head()

df.to_csv("cleaned_processes.csv", index=False)

## Analysis

For this analysis, it is important to distinguish between the different variables:
- The "item" variable refers to the object of each procurement process. It includes a unique combination of the active ingredient, pharmaceutical form, concentration, and commercial presentation.
- The variables 'principio_activo', 'forma_farmaceutica', 'concentracion','presentacion_comercial' are derived from "item", but will be used separately for the purposes of the analysis below.

#How many purchsings processes have been launched since 2022?
len(df)

#How many items have had at least one process launched
df["item"].nunique()

# How many different active principles had at least one process launched
df["principio_activo"].nunique()

principios = df["principio_activo"].unique().tolist()


#Which medicines have more than 1 process launched?
df["item"].value_counts().head().to_frame()

#Check the processes for Misoprostol launched 3 times
df[df["item"] == "DCI: MISOPROSTOL - FORMA FARMACÉUTICA: SÓLIDO ORAL - CONCENTRACIÓN: 200 MCG - PRESENTACIÓN COMERCIAL: CAJA X BLÍSTER/RISTRA"]

#Check the processes for Fenitoína launched 3 times
df[df["item"] == "DCI: FENITOÍNA - FORMA FARMACÉUTICA: SÓLIDO ORAL - CONCENTRACIÓN: 100 MG - PRESENTACIÓN COMERCIAL: CAJA X BLÍSTER/RISTRA/FRASCO DOSIS PERSONAL, NO ENVASE HOSPITALARIO"]

# Which kind of status of processes
df["estado"].value_counts().to_frame()

# Awarded contracts
len(df[df["estado"].str.contains("Adjudicado")])

# Failed contracts
len(df[~df["estado"].str.contains("Adjudicado")])

df["descripcion"].value_counts().to_frame()

------

### Items that got a contract

procesos_adjudicados = df[(df["estado"] == "Adjudicado oferente ganador") | (df["estado"] == "Adjudicado otros oferentes")]
len(procesos_adjudicados)

procesos_adjudicados.head(2)

procesos_adjudicados["principio_activo"].nunique()

procesos_adjudicados["item"].nunique()

procesos_adjudicados["principio_activo"].value_counts().to_frame().head()

procesos_adjudicados["descripcion"].value_counts()

procesos_adjudicados[procesos_adjudicados["principio_activo"] == "paracetamol"]

procesos_adjudicados.sort_values(by="presupuesto", ascending=False).head()

#Export so we can scrape the contract information
procesos_adjudicados.to_csv("adjudicados.csv", index=False)

---------

### Medicine that never got a contract
From the previous analysis, we can see there are items that never got a contract, even if they have several purchasing processes launched.

#Items that have an assigned supplier (adjudicado)
items_adjudicados = (df.groupby("item")["estado"]
    .apply(lambda x: x.isin(["Adjudicado oferente ganador", "Adjudicado otros oferentes"])
    .any())
              )

items_adjudicados.value_counts()

# Items that never got a contract
items_no_adjudicados = items_adjudicados[~items_adjudicados].index.to_list()
items_no_adjudicados = pd.DataFrame(items_no_adjudicados, columns=["item"])

len(items_no_adjudicados)

items_no_adjudicados["principio_activo"] = (items_no_adjudicados["item"]
    .str.split("DCI: ").str[1]
    .str.split("- FOR").str[0]
    .str.lower()
    .str.strip()
                                    )

items_no_adjudicados["forma_farmaceutica"] = (items_no_adjudicados["item"]
    .str.split("TICA:").str[1]
    .str.split("- CON").str[0]
    .str.lower()
    .str.strip()
                                      )

items_no_adjudicados["concentracion"] = (items_no_adjudicados["item"]
    .str.split("TRACIÓN:").str[1]
    .str.split("- PRE").str[0]
    .str.lower()
    .str.strip()
                                 )

items_no_adjudicados["presentacion_comercial"] = (items_no_adjudicados["item"]
    .str.split("CIAL:").str[1]
    .str.lower()
    .str.strip()
                                          )

items_no_adjudicados = pd.merge(
    items_no_adjudicados, 
    portfolio_clean[["NOMBRE GENERICO", "DESCRIPCIÓN DEL GRUPO ANATÓMICO"]], 
    left_on="principio_activo", 
    right_on="NOMBRE GENERICO", 
    how="left"
).drop(columns=["NOMBRE GENERICO"])

items_no_adjudicados = items_no_adjudicados.rename(columns={"DESCRIPCIÓN DEL GRUPO ANATÓMICO": "descripcion"})

items_no_adjudicados.head(2)

# Active principles that have failed processes the most
items_no_adjudicados["principio_activo"].value_counts().head().to_frame()

items_no_adjudicados["descripcion"].value_counts().to_frame()

#Example - what kinds of morfine didn't get a contract
items_no_adjudicados[items_no_adjudicados["item"].str.contains("MORFINA")]

# Is there morfine that DID get a contract?
df[(df["principio_activo"] == "morfina") & (df["estado"].str.contains("Adjudicado"))]

items_no_adjudicados.columns

# Is there active principles that, in any form, didn't get any contract?
adjudicados_principles = (df.groupby("principio_activo")["estado"]
    .apply(lambda x: x.isin(["Adjudicado oferente ganador", "Adjudicado otros oferentes"])
    .any())
                         )

adjudicados_principles.value_counts()

never_adjudicados_principles = adjudicados_principles[~adjudicados_principles].index.to_list()
len(never_adjudicados_principles)

items_no_adjudicados["tiene_otro_contrato"] = items_no_adjudicados["principio_activo"].apply(lambda x: "No" if x in never_adjudicados_principles else "Si")

items_no_adjudicados.to_csv("items_sin_contrato.csv", index=False)

-----

### Details for medicine that never got a contract

no_contract_principles = pd.DataFrame(never_adjudicados_principles, columns=["principio_activo"])
len(no_contract_principles)

no_contract_principles = pd.merge(
    no_contract_principles,
    df,
    on="principio_activo",
    how="left"
)

no_contract_principles["estado"].value_counts().to_frame()

no_contract_principles["descripcion"].value_counts().to_frame()

no_contract_principles["descripcion"].isna().value_counts()

no_contract_principles[no_contract_principles["descripcion"].isna()]

len(no_contract_principles)

no_contract_principles["principio_activo"].nunique()

no_contract_principles["item"].nunique()

no_contract_principles.head()

no_contract_principles.to_csv("principios_sin_contrato.csv", index=False)
