==> 3_procesos_adjudicados.ipynb <==
# Procesos adjudicados
This notebook will include the analysis of all the processes that got a concract. It has a scraper for getting extra information on each process, then uses merging for having other extra data about the medicine, and then includes the analysis. 

Each row in this analysis (414) is one item: one specific combination of a medicine, in a form, concentration and comercial presentation.

## 1. Scraper - details of each contract
Firstly, to have more info on each contract, we have a scraper for getting the info on quantity, item budget and complete budget.

from playwright.async_api import async_playwright
from bs4 import BeautifulSoup
import pandas as pd
import asyncio

df = pd.read_csv("adjudicados.csv")

df.head(1)

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=True)
page = await browser.new_page()

data = []

for i, row in df.head(5).iterrows():
    url = row["detalle_url"]
    item = {}
    item["codigo"] = row["codigo"]
    
    await asyncio.sleep(1)
    print(f"\rProcessing index: {i}, {url}", end="")
    try:
        max_attempts = 2
        for attempt in range(max_attempts):
            await page.goto(url, timeout=10000)
            await asyncio.sleep(3)
        
        try:
            await page.locator("a:has-text('Ver Resultados')").click()
            await asyncio.sleep(3)
        except:
            print(f"There is no 'Ver Resultados' in {url}")
            
        html = await page.content()
        soup = BeautifulSoup(html, "lxml")
        
        table = soup.select_one("table#rounded-corner tbody")
        row = table.find_all("td")
        if len(row) < 8:
            print(f"----- Incomplete row on {url}")
            continue
    
        item["cum_id"] = row[0].get_text(strip=True)
        item["cantidad"] = row[5].get_text(strip=True)
        item["precio_referencial"] = row[6].get_text(strip=True)
        item["subtotal"] = row[7].get_text(strip=True)


        adju_table = soup.select_one("table#tableadjudicacion")
        adju_rows = adju_table.find_all("tr")
        try:
            item["proveedor_adjudicado"] = adju_rows[0].find_all("td")[1].get_text(strip=True)
            item["valor_adjudicado"] = adju_rows[1].find_all("td")[1].get_text(strip=True)
            item["fecha_adjudicacion"] = adju_rows[3].find_all("td")[1].get_text(strip=True)
        except:
            continue
        
        data.append(item)    
        
    except:
        print(f"ERROR processing {url}")
        continue

len(data)

details = pd.json_normalize(data)

details.head()

details.info()

#I should clean details first, specially in numbers columns

df = pd.merge(df, details, on="codigo", how="left")

df.head(1)

len(df)

df.head(1)

summary.columns

summary = summary.rename(columns={"cantidad": "sum_cantidad_ordenes", "monto_total_adjudicado": "sum_valor_ordenes"})

merged = df.merge(summary, left_on="objeto", right_on="objeto_contractual")

merged = merged.drop("objeto_contractual", axis=1)

merged.head()

#Check if there are processes with no supplier...
merged[merged["proveedor"].isna()]

for url in merged[merged["proveedor"].isna()]["url"]:
    print(url)
