==> 4_scraper_contratos.ipynb <==
# Scraper para los pdfs de los convenios

from playwright.async_api import async_playwright
from bs4 import BeautifulSoup
import pandas as pd
import asyncio
import os

df = pd.read_csv("data/raw/complete_scraped_data.csv")

adjudicados = df[df["estado"].str.contains("Adjudicado")].reset_index()

len(adjudicados) #just the processes that have a contract

adjudicados.head()

os.makedirs('convenios/', exist_ok=True)

## Build the scraper

url = adjudicados["detalle_url"][1]
user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36"

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=False)
page = await browser.new_page(user_agent=user_agent)

url = adjudicados["detalle_url"][1]
await page.goto(url)
await asyncio.sleep(3)
await page.goto(url)

for i, row in adjudicados.iterrows():
    
    url = row["detalle_url"]
    code = row["codigo"]

    filename = f"{code}.pdf"
    save_path = f"convenios/{filename}"

    await asyncio.sleep(1)
    print(f"\rProcessing index: {i}, {url}", end="")
        
    await page.goto(url, timeout=10000)
    await asyncio.sleep(2)
            
    await page.wait_for_selector("#tab4", state="visible", timeout=10000)
    await page.eval_on_selector("#tab4", "el => el.click()")
    await asyncio.sleep(1)
    
    html = await page.content()
    soup = BeautifulSoup(html, "lxml")

    try:
        download_link_locator = page.locator(
            "//tr[translate(normalize-space(td[1]), 'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz')='convenio marco']//a[contains(@href, 'bajarArchivo.cpe')]"
        )
        assert await download_link_locator.count() == 1, "Expected exactly one CONVENIO MARCO link."
        
        async with page.expect_download() as download_info:
            await download_link_locator.click()

        download = await download_info.value
        await download.save_as(save_path)

        adjudicados.at[i, "contrato"] = "SI"
    
    except:
        print(f"Error at index {i}, {url}")
        adjudicados.at[i, "contrato"] = "NO"

os.makedirs('data/process', exist_ok=True)

adjudicados.to_csv("data/process/adjudicados_contratos.csv", index=False)

len(adjudicados[adjudicados["contrato"] == "NO"])

adjudicados[adjudicados["contrato"] == "NO"]["detalle_url"][36] #check one failed process

context = await browser.new_context()

sample_urls = adjudicados[adjudicados["contrato"] == "NO"]["detalle_url"].sample(10).tolist()

pages = []

for url in sample_urls:
    page = await context.new_page()
    await page.goto(url)
    pages.append(page)
    print(f"Opened {url}")

print(f"Opened {len(pages)} tabs. You can check them now.")

await asyncio.sleep(120)

#Now let's correct and loop thru the 44 with no contracts
for i, row in adjudicados[adjudicados["contrato"] == "NO"].iterrows():
    
    url = row["detalle_url"]
    code = row["codigo"]

    filename = f"{code}.pdf"
    save_path = f"convenios/{filename}"

    await asyncio.sleep(1)
    print(f"\rProcessing index: {i}, {url}", end="")
        
    await page.goto(url, timeout=10000)
    await asyncio.sleep(2)
            
    await page.wait_for_selector("#tab4", state="visible", timeout=10000)
    await page.eval_on_selector("#tab4", "el => el.click()")
    await asyncio.sleep(1)
    
    html = await page.content()
    soup = BeautifulSoup(html, "lxml")

    try:
        download_link_locator = page.locator(
            "//tr[starts-with(translate(normalize-space(td[1]), 'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz'), 'convenio marco')]//a[contains(@href, 'bajarArchivo.cpe')]"
        )
        assert await download_link_locator.count() == 1, "Expected exactly one CONVENIO MARCO link."
        
        async with page.expect_download() as download_info:
            await download_link_locator.click()

        download = await download_info.value
        await download.save_as(save_path)

        adjudicados.at[i, "contrato"] = "SI"
    
    except:
        print(f"Error at index {i}, {url}")
        adjudicados.at[i, "contrato"] = "NO"

len(adjudicados[adjudicados["contrato"] == "NO"])

adjudicados.to_csv("data/process/adjudicados_contratos.csv", index=False)

context = await browser.new_context()

sample_urls = adjudicados[adjudicados["contrato"] == "NO"]["detalle_url"].tolist()

pages = []

for url in sample_urls:
    page = await context.new_page()
    await page.goto(url)
    pages.append(page)
    print(f"Opened {url}")

print(f"Opened {len(pages)} tabs. You can check them now.")

await asyncio.sleep(120)

playwright = await browser.close()
